Updated LazyComponent import for Footer, ThemeSwitcher, and LazyLayout
Changed the import of LazyComponent in the Footer.php, ThemeSwitcher.php and LazyLayout.php files to use the Step2dev\LazyUI namespace instead of the Lazyadm\LazyComponent. This update aligns with the move of the LazyComponent class to the Step2dev\LazyUI package for cleaner package structure and easier maintainability.