Upgrade livewire dependency & refactor LazyAdminServiceProvider
Upgraded the "livewire/livewire" dependency from the beta version to stable version "^2.12|^3.0". This is to ensure the application stability. Also, commented out the call that publishes 'LazyAsideServiceProvider' from the LazyAdminServiceProvider. This was unnecessary and causing confusion. Added 'lazy-admin' to 'publish' to facilitate the LazyAdminServiceProvider's discoverability. PackageRegistered() was further improved to check the existence of Livewire and add essential macros to the Livewire component. These changes will significantly improve the performance and readability of the application.