The check for Livewire class and related macros was removed.
The check for the existence of 'Livewire\Livewire' and its related macros 'notify' and 'notifyFlash' in the 'packageRegistered' method of 'LazyAdminServiceProvider' was removed. This was due to the functionality being moved and handled differently in a newer implementation, making this check irrelevant and unnecessary. The removal leads to cleaner and more maintainable code.