Refactor Router class to use ConfigProvider
The Router class was updated to depend on a new ConfigProvider class. This was done to better manage the retrieval of admin route configuration parameters which were previously scattered throughout the Router class. This change greatly improves the organization and readability of the Router code. It also helps to keep up with best practices of class responsibility division.