Add menu rendering to layout
The layout.blade.php file has been updated to include a new parameter to handle the menu. Changes to the menu rendering logic in the `MenuManager` class were also made, changing the method name from `generate` to `render` for more context specificity. If the `menu` value is not specified in the layout, the `Menu::render()` method will be invoked by default.